+++ src/test/java/roomescape/reservation/waiting/controller/WaitingControllerTest.java
@@ -1,16 +1,14 @@
-package roomescape.reservation.controller;
+package roomescape.reservation.waiting.controller;
 
-import static org.hamcrest.Matchers.is;
-import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.any;
+import static org.mockito.Mockito.doNothing;
 
 import java.time.LocalDate;
 import java.util.Map;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.http.HttpStatus;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.annotation.DirtiesContext.ClassMode;
 import org.springframework.test.context.ActiveProfiles;
@@ -21,77 +19,22 @@ import io.restassured.http.ContentType;
 import roomescape.reservation.payment.dto.request.PaymentRequest;
 import roomescape.reservation.payment.service.PaymentService;
 
+//@Disabled
 @ActiveProfiles("test")
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = ClassMode.AFTER_EACH_TEST_METHOD)
-class AdminWaitingControllerTest {
+class WaitingControllerTest {
 
     @MockitoBean
     private PaymentService paymentService;
 
-    @DisplayName("모든 예약대기 목록을 읽어온다.")
+    @DisplayName("예약 대기를 추가한다.")
     @Test
-    void readAllWaiting() {
+    void createReservationWaiting() {
+        int timeId = addReservationTime("10:00");
+        int themeId = addTheme();
         String tokenValue = getAdminLoginTokenValue();
-        int timeId = addReservationTime("10:00");
-        int themeId = addTheme();
-        addReservation(tokenValue);
-
-        String memberLoginTokenValue = getMemberLoginTokenValue();
-        RestAssured.given()
-                .cookie("token", memberLoginTokenValue)
-                .contentType(ContentType.JSON)
-                .body(Map.of(
-                        "date", LocalDate.now().plusDays(1L),
-                        "timeId", timeId,
-                        "themeId", themeId
-                )).when().post("/waitings")
-                .then();
-
-        RestAssured.given().log().all()
-                .cookie("token", tokenValue)
-                .when().get("/admin/waitings")
-                .then().log().all()
-                .statusCode(200)
-                .body("size()", is(1));
-    }
-
-    @DisplayName("일반 사용자는 모든 예약대기 목록을 읽어올 수 없다.")
-    @Test
-    void readAllWaitingByMember() {
-        RestAssured.given()
-                .contentType(ContentType.JSON)
-                .body(Map.of(
-                        "email", "member@woowa.com",
-                        "password", "12341234",
-                        "name", "일반"
-                )).when().post("/members")
-                .then()
-                .statusCode(HttpStatus.CREATED.value());
-
-        String tokenValue = RestAssured.given()
-                .contentType(ContentType.JSON)
-                .body(Map.of(
-                        "email", "member@woowa.com",
-                        "password", "12341234"
-                )).when().post("/login")
-                .then()
-                .extract().cookie("token");
-
-        RestAssured.given().log().all()
-                .cookie("token", tokenValue)
-                .when().get("/admin/waitings")
-                .then().log().all()
-                .statusCode(401);
-    }
-
-    @DisplayName("예약대기를 거절한다.")
-    @Test
-    void denyWaiting() {
-        int timeId = addReservationTime("10:00");
-        int themeId = addTheme();
-        String adminLoginTokenValue = getAdminLoginTokenValue();
-        Map<String, Object> params = Map.of(
+        Map<String, Object> waitingParams = Map.of(
                 "date", getTomorrow(),
                 "timeId", timeId,
                 "themeId", themeId,
@@ -100,21 +43,73 @@ class AdminWaitingControllerTest {
                 "amount", 1_000L
         );
 
-        addReservation(adminLoginTokenValue);
+        doNothing().when(paymentService)
+                .confirm(any(PaymentRequest.class));
 
-        String memberLoginTokenValue = getMemberLoginTokenValue();
-        int waitingId = RestAssured.given()
+        RestAssured.given()
+                .cookie("token", tokenValue)
                 .contentType(ContentType.JSON)
-                .cookie("token", memberLoginTokenValue)
-                .body(params)
+                .body(waitingParams)
+                .when().post("/reservations")
+                .then();
+
+        String userLoginTokenValue = getUserLoginTokenValue();
+        RestAssured.given().log().all()
+                .cookie("token", userLoginTokenValue)
+                .contentType(ContentType.JSON)
+                .body(waitingParams)
+                .when().post("/waitings")
+                .then().log().all()
+                .statusCode(201);
+    }
+
+    @DisplayName("예약대기를 삭제한다.")
+    @Test
+    void deleteReservationWaiting() {
+        int timeId = addReservationTime("10:00");
+        int themeId = addTheme();
+        String tokenValue = getAdminLoginTokenValue();
+        Map<String, Object> reservationParams = Map.of(
+                "date", getTomorrow(),
+                "timeId", timeId,
+                "themeId", themeId,
+                "paymentKey", "paymentKey",
+                "orderId", "orderId",
+                "amount", 1_000L
+        );
+
+        doNothing().when(paymentService)
+                .confirm(any(PaymentRequest.class));
+
+        RestAssured.given()
+                .cookie("token", tokenValue)
+                .contentType(ContentType.JSON)
+                .body(reservationParams)
+                .when().post("/reservations")
+                .then();
+
+        String userLoginTokenValue = getUserLoginTokenValue();
+        int waitingId = RestAssured.given()
+                .cookie("token", userLoginTokenValue)
+                .contentType(ContentType.JSON)
+                .body(reservationParams)
                 .when().post("/waitings")
                 .then().extract().path("id");
 
         RestAssured.given().log().all()
-                .cookie("token", adminLoginTokenValue)
-                .when().delete("/admin/waitings/" + waitingId)
+                .cookie("token", userLoginTokenValue)
+                .when().delete("/waitings/" + waitingId)
                 .then().log().all()
                 .statusCode(204);
+    }
+
+    @DisplayName("존재하지 않는 예약대기를 삭제할 경우 NOT_FOUND 반환한다.")
+    @Test
+    void deleteNonExistsReservationWaiting() {
+        RestAssured.given().log().all()
+                .when().delete("/waitings/0")
+                .then().log().all()
+                .statusCode(404);
     }
 
     private String getAdminLoginTokenValue() {
@@ -127,7 +122,7 @@ class AdminWaitingControllerTest {
                 .extract().cookie("token");
     }
 
-    private String getMemberLoginTokenValue() {
+    private String getUserLoginTokenValue() {
         Map<String, String> adminLoginParams = Map.of("email", "user@woowa.com", "password", "12341234");
         return RestAssured.given()
                 .contentType(ContentType.JSON)
@@ -160,25 +155,5 @@ class AdminWaitingControllerTest {
                 .body(themeParams)
                 .when().post("/themes")
                 .then().extract().path("id");
-    }
-
-    private void addReservation(final String tokenValue) {
-        Mockito.doNothing()
-                .when(paymentService)
-                .confirm(any(PaymentRequest.class));
-
-        RestAssured.given()
-                .cookie("token", tokenValue)
-                .contentType(ContentType.JSON)
-                .body(Map.of(
-                        "date", LocalDate.now().plusDays(1L),
-                        "timeId", 1,
-                        "themeId", 1,
-                        "paymentKey", "paymentKey",
-                        "orderId", "orderId",
-                        "amount", 1_000L
-                ))
-                .when().post("/reservations")
-                .then();
     }
 }
